+++ R/RcppExports.R
@@ -1445,10 +1445,6 @@ Node___FF16__FF16_Env__compute_competition <- function(obj_, height) {
     .Call('_plant_Node___FF16__FF16_Env__compute_competition', PACKAGE = 'plant', obj_, height)
 }
 
-Node___FF16__FF16_Env__growth_rate_gradient <- function(obj_, environment) {
-    .Call('_plant_Node___FF16__FF16_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, environment)
-}
-
 Node___FF16__FF16_Env__compute_rates <- function(obj_, environment, pr_patch_survival) {
     invisible(.Call('_plant_Node___FF16__FF16_Env__compute_rates', PACKAGE = 'plant', obj_, environment, pr_patch_survival))
 }
@@ -1499,10 +1495,6 @@ Node___TF24__TF24_Env__ctor <- function(strategy) {
 
 Node___TF24__TF24_Env__compute_competition <- function(obj_, height) {
     .Call('_plant_Node___TF24__TF24_Env__compute_competition', PACKAGE = 'plant', obj_, height)
-}
-
-Node___TF24__TF24_Env__growth_rate_gradient <- function(obj_, environment) {
-    .Call('_plant_Node___TF24__TF24_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, environment)
 }
 
 Node___TF24__TF24_Env__compute_rates <- function(obj_, environment, pr_patch_survival) {
@@ -1557,10 +1549,6 @@ Node___TF24f__TF24_Env__compute_competition <- function(obj_, height) {
     .Call('_plant_Node___TF24f__TF24_Env__compute_competition', PACKAGE = 'plant', obj_, height)
 }
 
-Node___TF24f__TF24_Env__growth_rate_gradient <- function(obj_, environment) {
-    .Call('_plant_Node___TF24f__TF24_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, environment)
-}
-
 Node___TF24f__TF24_Env__compute_rates <- function(obj_, environment, pr_patch_survival) {
     invisible(.Call('_plant_Node___TF24f__TF24_Env__compute_rates', PACKAGE = 'plant', obj_, environment, pr_patch_survival))
 }
@@ -1611,10 +1599,6 @@ Node___K93__K93_Env__ctor <- function(strategy) {
 
 Node___K93__K93_Env__compute_competition <- function(obj_, height) {
     .Call('_plant_Node___K93__K93_Env__compute_competition', PACKAGE = 'plant', obj_, height)
-}
-
-Node___K93__K93_Env__growth_rate_gradient <- function(obj_, environment) {
-    .Call('_plant_Node___K93__K93_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, environment)
 }
 
 Node___K93__K93_Env__compute_rates <- function(obj_, environment, pr_patch_survival) {
@@ -1683,6 +1667,10 @@ Species___FF16__FF16_Env__introduce_new_node <- function(obj_) {
 
 Species___FF16__FF16_Env__node_at <- function(obj_, index) {
     .Call('_plant_Species___FF16__FF16_Env__node_at', PACKAGE = 'plant', obj_, index)
+}
+
+Species___FF16__FF16_Env__growth_rate_gradient <- function(obj_, index) {
+    .Call('_plant_Species___FF16__FF16_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, index)
 }
 
 Species___FF16__FF16_Env__compute_competition_effect_by_nodes_error <- function(obj_, scal) {
@@ -1785,6 +1773,10 @@ Species___TF24__TF24_Env__node_at <- function(obj_, index) {
     .Call('_plant_Species___TF24__TF24_Env__node_at', PACKAGE = 'plant', obj_, index)
 }
 
+Species___TF24__TF24_Env__growth_rate_gradient <- function(obj_, index) {
+    .Call('_plant_Species___TF24__TF24_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, index)
+}
+
 Species___TF24__TF24_Env__compute_competition_effect_by_nodes_error <- function(obj_, scal) {
     .Call('_plant_Species___TF24__TF24_Env__compute_competition_effect_by_nodes_error', PACKAGE = 'plant', obj_, scal)
 }
@@ -1885,6 +1877,10 @@ Species___TF24f__TF24_Env__node_at <- function(obj_, index) {
     .Call('_plant_Species___TF24f__TF24_Env__node_at', PACKAGE = 'plant', obj_, index)
 }
 
+Species___TF24f__TF24_Env__growth_rate_gradient <- function(obj_, index) {
+    .Call('_plant_Species___TF24f__TF24_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, index)
+}
+
 Species___TF24f__TF24_Env__compute_competition_effect_by_nodes_error <- function(obj_, scal) {
     .Call('_plant_Species___TF24f__TF24_Env__compute_competition_effect_by_nodes_error', PACKAGE = 'plant', obj_, scal)
 }
@@ -1983,6 +1979,10 @@ Species___K93__K93_Env__introduce_new_node <- function(obj_) {
 
 Species___K93__K93_Env__node_at <- function(obj_, index) {
     .Call('_plant_Species___K93__K93_Env__node_at', PACKAGE = 'plant', obj_, index)
+}
+
+Species___K93__K93_Env__growth_rate_gradient <- function(obj_, index) {
+    .Call('_plant_Species___K93__K93_Env__growth_rate_gradient', PACKAGE = 'plant', obj_, index)
 }
 
 Species___K93__K93_Env__compute_competition_effect_by_nodes_error <- function(obj_, scal) {